change to searchIndexBuilder
